[FEAT] INT-565: Add initial UseCase support (#135)
## Problem
We want to provide new "use cases" in Console to help folks find the best workflow for their jobs. We want to be able to make quick changes to the available use cases, including associated thumbnail images, without needing to redeploy Console.

## Solution
Define the use case card configs here in gretel-blueprints repo, which we can then query from the Console. This also adds the initial set of images we plan to use. I added a test suite around the use cases, including using JSON Schema to validate that each card conforms to a data shape we are expecting (e.g., required fields, enum values). In a followup PR I plan to add a GitHub Action to run these automatically in CI. _Misc: updated repo README to include new directory info._

## Notes
Initial use cases / text is likely to change still, these are placeholder values that will let me keep developing and wire everything together. The great thing is that the way we're designing this makes it easy to update these!