Add filesize (bytes) to sample datasets referenced in use cases (#146)
## Problem
We frequently want to display the file size of sample datasets to folks viewing them in Console. When we initially added them to the Use Cases JSON, we forgot to add this detail!

## Solution
Add this data to the sample dataset metadata within the use cases. This allows us to display the file size without needing to load the file in Console, and keeps all the important info about the file in one place. Since we already need to update this info any time a file changes (and they don't change frequently), this feels like a decent, simple solution for now.