[FIX] update use case model types, fix config file path for one (#148)
## Problem
When we initially built out the use cases JSON, we set the modeltype for amplify tasks as "amplify" but they're technically 'synthetics'. Console needs them to be marked as synthetics due to how the config templates are currently organized, too. Because it wasn't organized in this way, grabbing the correct config template for a use case wasn't working well. I also noticed that one use case specified a config file that didn't actually exist. oops!

## Solution
Update amplify use cases to specify a model type of synthetics. Fix the file name. Add a validation check that the config files for use cases actually exist.