[FEAT] Add sample data previews. New dir, script, test, and github action (#188)
## Problem
We want to display the first 20 records of sample datasets in some places on Console, so that folks can see at a glance what data they're working with. We want to avoid downloading the full datasets when we do this, and we don't want to hardcode the data into Console.

## Solution
Create a new entity, "sample data preview", that lives in its own folder. I've also created a script to generate these easily, and a test that runs in a GitHub action to verify that all sample data files have a corresponding preview file. This doesn't solve for _changing_ a sample dataset's contents, but that is pretty unusual today.

## Note
It would be nice to have these files get autogenerated for us -- maybe via GitHub action? Given we don't add/change the sample datasets much today, it felt like an extra step we can live without for now.