[FIX] Sample data previews need to be written as valid CSVs (#194)
## Problem
When I implemented the sample data preview generation, I tried to avoid using an extra library by doing the file-writing myself. I didn't realize this was making some files (particularly those with lots of text, newlines) in a way that couldn't be properly parsed back out.

## Solution
Use the `stringify` library that is the counterpart to the `parse` library that I've used for the CSV parsing. This library knows how to properly "unparse" the CSV.

## Testing
Check out the preview CSVs, particularly the taylor swift lyrics and sample_transform_emails. If you open them in a spreadsheet software, they should get imported correctly now.