2023-01-18 Release: Use Case Card Updates and Sample Dataset Previews (#198)
* [FEAT] Add sample data previews. New dir, script, test, and github action (#188)

We want to display the first 20 records of sample datasets in some places on Console, so that folks can see at a glance what data they're working with. We want to avoid downloading the full datasets when we do this, and we don't want to hardcode the data into Console.

Create a new entity, "sample data preview", that lives in its own folder. I've also created a script to generate these easily, and a test that runs in a GitHub action to verify that all sample data files have a corresponding preview file. This doesn't solve for _changing_ a sample dataset's contents, but that is pretty unusual today.

It would be nice to have these files get autogenerated for us -- maybe via GitHub action? Given we don't add/change the sample datasets much today, it felt like an extra step we can live without for now.

* [FIX] Change Icon for dimensional use case (#195)

* [FIX] Sample data previews need to be written as valid CSVs (#194)

## Problem
When I implemented the sample data preview generation, I tried to avoid using an extra library by doing the file-writing myself. I didn't realize this was making some files (particularly those with lots of text, newlines) in a way that couldn't be properly parsed back out.

## Solution
Use the `stringify` library that is the counterpart to the `parse` library that I've used for the CSV parsing. This library knows how to properly "unparse" the CSV.

## Testing
Check out the preview CSVs, particularly the taylor swift lyrics and sample_transform_emails. If you open them in a spreadsheet software, they should get imported correctly now.

* Add Benchmark notebook card (#196)

Add Benchmark notebook card and remove the Downstream ML accuracy one (for now)

* added downstream ML UCC back (#197)

We decided to keep it in after all

Co-authored-by: Yamini <yamini@users.noreply.github.com>