Support for Transform V2 (#324)
## Problem
We want to add support to Console for Transform V2. This requires quite a few pieces that are not currently in this repo.

## Solution
First, I reverted our develop-only change around the synthetic card, because we aren't ready for that to go to prod. It's easy to redo if we'd still like to test it out, but it's been sitting for a while with no action.

Then, I added the following **new** items:
* sample dataset with preview
* new config template
* new model type in the model types definition
* blueprint card image
* blueprint card definition, including a bit of reordering of cards.

And, updated the following:
* replace the transform v1 image per design request.

Also:
* Revert "[FEAT] Add markdown file and button placeholder for synth console card (#318)"
This reverts commit 36a3379ce8ba25fd5c8a50f4408efbbf1cb31943.