Add some relational samples (#347) (#348)
* Add some relational samples (#347)

# Problem

We want to be able to create blueprints that use sample connections provided to ease the onboarding process. This is required to create blueprints that use databases because all database workflows are powered by connections, as opposed to a sample static file.

# Solution

Create 2 new blueprints that use a new `sampleConnection` field and update unit tests to support this field. Also reorg a bit so there aren't duplicate blueprints and the homepage has the desired use cases