[RELEASE] 2024-02-23 prompts and usecase images (#353)
* [IDEA] Add Playground prompt examples to gretel blueprints repo. (#352)

# Problem
Playground prompt examples come to us from product. When we initially built out playground, they changed frequently, and every update required a code change and deploy. This took up engineering time that could have been used for other things. Example prompts change less frequently now, but we still update them occasionally. If we launch new capabilities on Playground, we'll likely have some iteration on them again.

# Solution
Move the prompt examples to this repo, where product can more easily adjust them with less engineer involvement (increased efficiency -- a win!). We can also iterate all we want in development environments here, and only move examples to production when we are ready to do so. No feature flag shenanigans required!

Single line prompts can be _just a string_. JSON doesn't support multi-line strings in a very human readable way, so the prompt `value` can also be an array of strings, with each new line a separate string. This formats nicely for human reading and editing, and is easy enough to parse in Console.

# Notes
This was an idea Stefan and I have been kicking around for a while. Totally open to adjusting things here if something is easier for product folks to edit. Ideally, I would like to keep these in one file for now for ease of fetching on the Console side.

* clean up use case images