[FEAT] INT-1898 Update JS Navigator script to allow > 50 rows (#388)
## Problem
We have an example script in JavaScript for using the Navigator APIs to generate data. It only enables the ability to create one stream at a time, therefore limiting the user to a maximum of 50 rows of data fetching per call. 

## Solution
We've written some Python code to facilitate calling more than 50 rows, so let's do it for JS, too. This diff updates the JS "sdk" to create multiple streams sequentially as needed to fulfill the requested number of rows. This includes passing the results of the previous stream into the new stream to facilitate data generation.

## Additional Info
I refactored the existing code a bit to pull out some sub-functions and generally tidy a few things. I also added the ability to output the data as json rather than as the table, since users might want to copy the data for use elsewhere. Try `--json=true` for that functionality.

## Testing
Pull this down locally. Get your API key (note: either use prod environment or update the `BASE_URL` to the dev environment). To see it create and use multiple streams, pass a `num_rows` > 50, such as:

```
GRETEL_API_KEY={your key} node example.js --prompt="generate a table of cat names and their favorite toys" --num_rows=52
```